Overhaul the NameValidation class.
Makes all of the individual tests private, resulting in the classes that
needed them instead calling specific Name Validation methods which are
built for their object type.

Some inconsistencies were uncovered:
- we're now using the config-blacklisted-names lists more thoroughly,
which should result in less shenanigans in titles and boards.
- we're also now making towns/nations names play by more consistent
rules, I believe I find some oversights which are now closed.

TODO: make sure that the new InvalidNameException messages are being passed on to the user. This might require updating the exception to behave more like a TownyException or dropping them entirely for TownyExceptions.